# docs/superadmin-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/superadmin/tenants": {
      "get": {
        "operationId": "TenantsController_listTenants",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name or email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE"
              ]
            }
          },
          {
            "name": "subscriptionStatus",
            "required": false,
            "in": "query",
            "description": "Filter by subscription status",
            "schema": {
              "type": "string",
              "enum": [
                "TRIAL",
                "ACTIVE",
                "PAST_DUE",
                "CANCELED"
              ]
            }
          },
          {
            "name": "planCode",
            "required": false,
            "in": "query",
            "description": "Filter by plan code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isOnboarded",
            "required": false,
            "in": "query",
            "description": "Filter by onboarding status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated tenant collection",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "tenant_123",
                        "name": "Acme Corp",
                        "email": "owner@acme.com",
                        "status": "ACTIVE",
                        "isOnboarded": true,
                        "onboardedAt": "2025-05-02T12:00:00.000Z",
                        "createdAt": "2025-05-01T12:34:56.000Z",
                        "updatedAt": "2025-05-02T12:34:56.000Z",
                        "lastActiveAt": "2025-05-04T09:15:00.000Z"
                      }
                    ],
                    "meta": {
                      "page": 1,
                      "pageSize": 20,
                      "total": 42,
                      "totalPages": 3
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List tenants with filters and pagination",
        "tags": [
          "Tenants"
        ]
      },
      "post": {
        "operationId": "TenantsController_createTenant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created tenant record",
            "schema": {
              "example": {
                "id": "tenant_456",
                "name": "Nova LLC",
                "email": "founder@nova.com",
                "status": "ACTIVE",
                "createdAt": "2025-05-10T08:00:00.000Z",
                "updatedAt": "2025-05-10T08:00:00.000Z",
                "lastActiveAt": null
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create tenant",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/api/v1/superadmin/tenants/{id}": {
      "get": {
        "operationId": "TenantsController_getTenant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get tenant details",
        "tags": [
          "Tenants"
        ]
      },
      "patch": {
        "operationId": "TenantsController_updateTenant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update tenant details",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/api/v1/superadmin/tenants/{id}/activate": {
      "patch": {
        "operationId": "TenantsController_activateTenant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Activate tenant",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/api/v1/superadmin/tenants/{id}/deactivate": {
      "patch": {
        "operationId": "TenantsController_deactivateTenant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Deactivate tenant",
        "tags": [
          "Tenants"
        ]
      }
    },
...
    "/api/v1/superadmin/tenants/{id}/stats": {
      "get": {
        "operationId": "TenantsController_getTenantStats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant stats payload",
            "schema": {
              "example": {
                "tenantId": "tenant_123",
                "tenantName": "Acme Corp",
                "status": "ACTIVE",
                "lastActiveAt": "2025-05-01T12:34:56.000Z",
                "metricsLogged": 120,
                "outcomesCompleted": 18,
                "activitiesLogged": 45,
                "salesLogged": 9,
                "momentumScore": 78,
                "streak": 6
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantStatsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get tenant usage stats",
        "tags": [
          "Tenants"
        ]
      }
    },
...
    "/api/v1/superadmin/reports/tenant/{id}/summary": {
      "get": {
        "operationId": "ReportsController_tenantSummary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for report range (ISO date)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for report range (ISO date)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantReportDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Generate tenant summary report",
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/superadmin/reports/tenant/{id}/export": {
      "get": {
        "operationId": "ReportsController_exportTenantReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for report range (ISO date)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for report range (ISO date)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file download"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Export tenant report as CSV",
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/superadmin/reports/platform/summary": {
      "get": {
        "operationId": "ReportsController_platformSummary",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for report range (ISO date)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for report range (ISO date)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Platform KPIs and charts data",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "kpis": {
                      "totalTenants": 42,
                      "activeTenants": 30,
                      "totalRevenue": 125000,
                      "averageMomentum": 72
                    },
                    "charts": {
                      "tenantGrowth": [],
                      "revenueByPlan": [],
                      "activityTrend": []
                    },
                    "generatedAt": "2025-06-01T12:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Generate platform-wide summary report",
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/superadmin/reports/platform/export": {
      "get": {
        "operationId": "ReportsController_exportPlatformReport",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for report range (ISO date)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for report range (ISO date)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file download"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Export platform report as CSV",
        "tags": [
          "Reports"
        ]
      }
    }
  },
  "info": {
    "title": "BG Accountability SuperAdmin API",
    "description": "REST API for the BG Accountability Platform - SuperAdmin Services.\n\n## Overview\nThis API provides administrative capabilities for managing the multi-tenant platform.\nAccess is restricted to users with the `SUPER_ADMIN` role.\n\n## Authentication\nAll endpoints (except login) require a valid JWT Bearer token.\n- Login via `POST /api/v1/auth/login`\n- Include token in header: `Authorization: Bearer <token>`\n- Refresh tokens via `POST /api/v1/auth/refresh`\n\n## Multi-Factor Authentication (MFA)\nSuperAdmin accounts require MFA for enhanced security.\n- If MFA is enabled, login returns a `tempToken`\n- Complete MFA via `POST /api/v1/auth/mfa/login` with tempToken + TOTP code\n- MFA can be managed via `/api/v1/auth/mfa/*` endpoints\n\n## Session Management\n- View active sessions: `GET /api/v1/auth/sessions`\n- Revoke sessions: `DELETE /api/v1/auth/sessions/:id`\n\n## Rate Limiting\n- Default: 100 requests per minute per IP\n- Auth endpoints: 10 requests per minute per IP\n\n## Error Format\nAll errors follow: `{ success: false, message: string, code?: string, details?: object }`\n\n## Pagination\nPaginated endpoints use: `{ items: T[], total: number, page: number, pageSize: number }`\n",
    "version": "1.0.0",
    "contact": {
      "name": "BG Accountability",
      "url": "https://bridgegaps.app",
      "email": "support@bridgegaps.app"
    },
    "license": {
...
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token (SUPER_ADMIN role required)"
      }
    },
    "schemas": {
...
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Company or freelancer name"
          },
          "email": {
            "type": "string",
            "description": "Primary contact email"
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "TenantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "tenant_123"
          },
          "name": {
            "type": "string",
            "example": "Acme Corp"
          },
          "email": {
            "type": "string",
            "example": "owner@acme.com"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          },
          "subscriptionStatus": {
            "type": "string",
            "enum": [
              "TRIAL",
              "ACTIVE",
              "PAST_DUE",
              "CANCELED"
            ],
            "example": "TRIAL"
          },
          "planCode": {
            "type": "object",
            "example": "GROWTH",
            "nullable": true
          },
          "renewalDate": {
            "type": "object",
            "example": "2025-07-01T00:00:00.000Z",
            "nullable": true
          },
          "trialEndsAt": {
            "type": "object",
            "example": "2025-06-15T00:00:00.000Z",
            "nullable": true
          },
          "isOnboarded": {
            "type": "boolean",
            "example": true,
            "description": "Whether the tenant has completed onboarding"
          },
          "onboardedAt": {
            "type": "object",
            "example": "2025-05-02T14:30:00.000Z",
            "nullable": true,
            "description": "When onboarding was completed"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-05-01T12:34:56.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-05-15T08:20:00.000Z"
          },
          "lastActiveAt": {
            "type": "object",
            "example": "2025-05-15T08:20:00.000Z",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "status",
          "subscriptionStatus",
          "planCode",
          "renewalDate",
          "trialEndsAt",
          "isOnboarded",
          "onboardedAt",
          "createdAt",
          "updatedAt",
          "lastActiveAt"
        ]
      },
      "UpdateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Company or freelancer name"
          },
          "email": {
            "type": "string",
            "description": "Primary contact email"
          }
        }
      },
...
      "TenantStatsDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "example": "tenant_123"
          },
          "tenantName": {
            "type": "string",
            "example": "Acme Corp"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          },
          "lastActiveAt": {
            "type": "object",
            "example": "2025-05-20T10:00:00.000Z",
            "nullable": true
          },
          "metricsLogged": {
            "type": "number",
            "example": 120
          },
          "outcomesCompleted": {
            "type": "number",
            "example": 18
          },
          "activitiesLogged": {
            "type": "number",
            "example": 45
          },
          "salesLogged": {
            "type": "number",
            "example": 9
          },
          "momentumScore": {
            "type": "number",
            "example": 78
          },
          "streak": {
            "type": "number",
            "example": 6
          }
        },
        "required": [
          "tenantId",
          "tenantName",
          "status",
          "lastActiveAt",
          "metricsLogged",
          "outcomesCompleted",
          "activitiesLogged",
          "salesLogged",
          "momentumScore",
          "streak"
        ]
      },
...
      "TenantReportTenantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "tenant_123"
          },
          "name": {
            "type": "string",
            "example": "Acme Corp"
          },
          "email": {
            "type": "string",
            "example": "owner@acme.com"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          },
          "subscriptionStatus": {
            "type": "string",
            "enum": [
              "TRIAL",
              "ACTIVE",
              "PAST_DUE",
              "CANCELED"
            ],
            "example": "ACTIVE"
          },
          "planCode": {
            "type": "object",
            "example": "GROWTH",
            "nullable": true
          },
          "renewalDate": {
            "type": "object",
            "example": "2025-07-01T00:00:00.000Z",
            "nullable": true
          },
          "trialEndsAt": {
            "type": "object",
            "example": "2025-06-15T00:00:00.000Z",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "status",
          "subscriptionStatus",
          "planCode",
          "renewalDate",
          "trialEndsAt"
        ]
      },
      "TenantReportUsageDto": {
        "type": "object",
        "properties": {
          "metricsLogged": {
            "type": "number",
            "example": 120
          },
          "outcomesCompleted": {
            "type": "number",
            "example": 15
          },
          "activitiesLogged": {
            "type": "number",
            "example": 42
          },
          "salesLogged": {
            "type": "number",
            "example": 12
          },
          "momentumScore": {
            "type": "number",
            "example": 78
          },
          "streak": {
            "type": "number",
            "example": 5
          },
          "lastActiveAt": {
            "type": "object",
            "example": "2025-06-02T00:00:00.000Z",
            "nullable": true
          }
        },
        "required": [
          "metricsLogged",
          "outcomesCompleted",
          "activitiesLogged",
          "salesLogged",
          "momentumScore",
          "streak",
          "lastActiveAt"
        ]
      },
      "TenantReportDto": {
        "type": "object",
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantReportTenantDto"
          },
          "usage": {
            "$ref": "#/components/schemas/TenantReportUsageDto"
          }
        },
        "required": [
          "tenant",
          "usage"
        ]
      }
    }
  }
}